chore: ndk버전 오류 해결 및 pubspec.yaml이미지 경로 세팅 완료
