--- app/lib/src/app.dart
@@ -1,25 +1,0 @@
-import 'package:flutter/material.dart';
-
-class App extends StatelessWidget {
-  const App({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    debugPrint(context.toString());
-    return MaterialApp(
-      /// App title.
-      title: 'Talk Pilot',
-      theme: ThemeData(
-        /// Default theme color
-        /// 추후 변경 예정
-        primarySwatch: Colors.blue,
-      ),
-
-      /// debug banner 삭제
-      debugShowCheckedModeBanner: false,
-
-      /// default home screen
-      home: const Scaffold(),
-    );
-  }
-}
