+++ app/pubspec.yaml
@@ -53,6 +53,8 @@ dev_dependencies:
 
 # The following section is specific to Flutter packages.
 flutter:
+  assets:
+    - assets/login/
 
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
